refactor(content): extract shortcode pipeline into modules
