+++ src/content/mod.rs
@@ -3,3 +3,4 @@ pub mod frontmatter;
 pub mod loader;
 pub mod markdown;
 pub mod pages;
+pub mod shortcodes;
